+++ src/controls/radio.ts
@@ -1,4 +1,3 @@
-import nestedProperty from "../helpers/nestedProperty";
 import { WUPcssHidden } from "../styles";
 import WUPBaseControl from "./baseControl";
 
@@ -309,14 +308,13 @@ export default class WUPRadioControl<
     if (isNeedRenderItems) {
       this.#cachedItems = undefined;
       // it's important to be before super otherwise initValue won't work
-      const attr = this.getAttribute("items");
-      this._opts.items = attr ? (nestedProperty.get(window, attr) as WUP.Radio.Options["items"]) : this._opts.items;
+      this._opts.items = this.getAttr<WUP.Radio.Options["items"]>("items", "ref") || [];
       this.renderItems(this.$refFieldset);
     }
 
     super.gotChanges(propsChanged as any);
 
-    this._opts.reverse = this.getBoolAttr("reverse", this._opts.reverse);
+    this._opts.reverse = this.getAttr("reverse", "bool");
     this.setAttr("reverse", this._opts.reverse, true);
 
     const req = this.validations?.required;